S3: fix bug when self.base == None
boto3.client and boto3.resource return different types of objects and
only the latter in the latter case does the returned instance have a
buckets attribute.
This patch uses boto3.resource regardless of self.base being set, which
works fine.
It will now also appropriately deal with IAM policies that do not allow
its user to retrieve the list of buckets, and will instead check
directly if the bucket named $SREGISTRY_S3_BUCKET exists or not.